Finish building out info tile display There are a TON of changes in this commit. Sorry for lumping everything together, but I set a goal to get a working preview of this bad boy out by the end of the day.
Primarily, this commit adds a new PropertiesTile module which represents a DOM element that actually displays properties of different chart objects. There were also some modifications to the DataSourceAdapter module to provide properties in the format we need them. I modified app.js to initialize the PropertiesTile and link it up to click events on the chart.

There is still a long way to go in the way of optimization. This should be ready to move to the release branch, but before it's finalized I have to add a ton of documentation and mobile optimize the styling.

We're getting close!